refactor: split components and utils into separate files
